move nav menu classes from Theme to Nav_Menus namespace
